Ugly workaround for a nasty problem with SQLite
The problem manifests itself in that some change rows are returned
more than once.

AFAICS, while we "SELECT FROM change" and loop over the result set
to generate changesets we "UPDATE change" to set the changeset_id
and we COMMIT after all updates for a new changeset are done.  That
is obviously a problem, because we UPDATE the table while selecting
from it.

Now the SQLite documentation says it is okay to use other tables
while we do a SELECT on one.  So that would mean we could create a
temporary table and select the changes from that so that we are
allowed to update the original table again, but in practice that
didn't change a thing!  As long as we still COMMIT after a change,
the problem stays. :-(