Fixed expansion of $Header$
OpenBSD's co(1) at least expands the RCS filename to the real path
with symlinks resolved but for the conversion to be perfect the path
has to be used as it was passed to co.

While there, provide better diagnostic message in case of an
assertion failure because unicode and binary strings were mixed.