Rewrite keyword expansion to fix $Log$
This also handles a few more keywords.  The main reason for this
change was that it looked as if RCS would expand *both* keywords
in a substring such as "$Log$Id$" and that was impossible to do
with a single-pass non-overlapping re substitution.  Unfortunately
I did not check the original RCS file and forgot which one caused
the problem, but I have doubts whether the original string was
really $Log$Id$.

Anyway, clone --verify should now work with OpenBSD's src repository.